requirements.txt
langchain>=0.3
langchain-openai>=0.2
langgraph>=0.2
langgraph-checkpoint-sqlite>=2.0
streamlit>=1.35
feedparser>=6.0
beautifulsoup4>=4.12
requests>=2.31
pyyaml>=6.0
python-dotenv>=1.0
aiohttp>=3.9
